Update pages api graphql context with user object and prisma client
